fix cc files for new matlab fix data path

+++ code_full/mex_unix/resize.cc
@@ -82,7 +82,7 @@ void resize1dtran(double *src, int sheight, double *dst, int dheight,
 // returns resized image
 mxArray *resize(const mxArray *mxsrc, const mxArray *mxscale) {
   double *src = (double *)mxGetPr(mxsrc);
-  const int *sdims = mxGetDimensions(mxsrc);
+  const long unsigned int *sdims = mxGetDimensions(mxsrc);
   if (mxGetNumberOfDimensions(mxsrc) != 3 || 
       mxGetClassID(mxsrc) != mxDOUBLE_CLASS)
     mexErrMsgTxt("Invalid input");  
@@ -91,7 +91,7 @@ mxArray *resize(const mxArray *mxsrc, const mxArray *mxscale) {
   if (scale > 1)
     mexErrMsgTxt("Invalid scaling factor");   
 
-  int ddims[3];
+  long unsigned int ddims[3];
   ddims[0] = (int)round(sdims[0]*scale);
   ddims[1] = (int)round(sdims[1]*scale);
   ddims[2] = sdims[2];
